requirements.txt
openpyxl>=3.1.0
PyYAML>=6.0
tk>=0.1.0
cryptography>=3.4.0
